BlueSkyPythonApp/BlueSkyPythonApp/requirements.txt
annotated-types==0.7.0
anyio==4.9.0
atproto==0.0.61
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.0
cryptography==44.0.3
dnspython==2.7.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
libipld==3.0.1
pillow==11.2.1
pycparser==2.22
pydantic==2.11.4
pydantic_core==2.33.2
requests==2.32.3
sniffio==1.3.1
typing-inspection==0.4.0
typing_extensions==4.13.2
urllib3==2.4.0
websockets==13.1
